+++ src/common/swc/utils.ts
@@ -1,11 +1,17 @@
 import path from 'path';
 import {convert} from 'tsconfig-to-swcconfig';
 
+const DEFAULT_EXCLUDE = ['node_modules'];
+
 export const EXTENSIONS_TO_COMPILE = ['.js', '.ts', '.mts', '.mjs', '.cjs'];
 
 function resolvePaths(paths: Record<string, string[]>, baseUrl: string) {
     const entries = [];
-    for (const targets of Object.values(paths)) {
+    for (const [key, targets] of Object.entries(paths)) {
+        if (key === '*') {
+            continue;
+        }
+
         for (const target of targets) {
             const resolvedPath = path.resolve(baseUrl, target.replace(/\*$/, ''));
             entries.push(resolvedPath);
@@ -14,17 +20,56 @@ function resolvePaths(paths: Record<string, string[]>, baseUrl: string) {
     return entries;
 }
 
-export function getSwcOptionsFromTsconfig(projectPath: string, filename = 'tsconfig.json') {
+export interface GetSwcOptionsParams {
+    projectPath: string;
+    filename?: string;
+    additionalPaths?: string[];
+    exclude?: string | string[];
+    publicPath: string;
+}
+
+export function getSwcOptions({
+    projectPath,
+    filename = 'tsconfig.json',
+    additionalPaths,
+    exclude,
+    publicPath,
+}: GetSwcOptionsParams) {
     const swcOptions = convert(filename, projectPath);
+    swcOptions.exclude = swcOptions.exclude || [];
     swcOptions.jsc = {
         ...swcOptions.jsc,
         // SWC requires absolute path as baseUrl
         baseUrl: projectPath,
+        transform: {
+            ...swcOptions.jsc?.transform,
+            optimizer: {
+                ...swcOptions.jsc?.transform?.optimizer,
+                globals: {
+                    ...swcOptions.jsc?.transform?.optimizer?.globals,
+                    vars: {
+                        'process.env.PUBLIC_PATH': JSON.stringify(publicPath),
+                        ...swcOptions.jsc?.transform?.optimizer?.globals?.vars,
+                    },
+                },
+            },
+        },
     };
+
+    let customExclude: string[] = [];
+    if (Array.isArray(exclude)) {
+        customExclude = exclude;
+    } else if (exclude) {
+        customExclude = [exclude];
+    }
+
+    swcOptions.exclude = [...(swcOptions.exclude || []), ...DEFAULT_EXCLUDE, ...customExclude];
 
     // SWC don't compile referenced files like tsc, so we need collect all directories to compile.
     const paths = swcOptions.jsc.paths || {};
-    const directoriesToCompile = [projectPath, ...resolvePaths(paths, projectPath)];
+    const directoriesToCompile = [
+        ...new Set([projectPath, ...resolvePaths(paths, projectPath), ...(additionalPaths || [])]),
+    ];
 
     return {
         swcOptions,